// src/app/modules/mapas/interfaces/country.interface.ts
export interface Country {

    name: any;
    tld: string[]; 
    cca2: string; 
    ccn3: string;
    cca3: string;
    cioc: string;
    independent: boolean;
    status: string;
    unMember: boolean;
    currencies: Currencies;
    idd: Idd;
    capital: string[];
    altSpellings: string[];
    region: string;
    subregion: string;  
    languages: Languages;  
    translations: { [key: string]: Translation };  
    latlng: number[];  
    landlocked: boolean;  
    borders: string[];  
    area: number;  
    demonyms: Demonyms;  
    flag: string;  
    maps: Maps;  
    population: number;  
    gini: Gini;  
    fifa: string;  
    car: Car;  
    timezones: string[];  
    continents: string[];  
    flags: CoatOfArms;  
    coatOfArms: CoatOfArms;  
    startOfWeek: string;  
    capitalInfo: CapitalInfo;  
    postalCode: PostalCode;  
  }
  
  
  
  export interface CapitalInfo {
  
    latlng: number[];
  
  }
  
  
  
  export interface Car {
  
    signs: string[];
  
    side: string;
  
  }
  
  
  
  export interface CoatOfArms {
  
    png: string;
  
    svg: string;
  
  }
  
  
  
  export interface Currencies {
  
    EUR: Eur;
  
  }
  
  
  
  export interface Eur {
  
    name: string;
  
    symbol: string;
  
  }
  
  
  
  export interface Demonyms {
  
    eng: Eng;
  
    fra: Eng;
  
  }
  
  
  
  export interface Eng {
  
    f: string;
  
    m: string;
  
  }
  
  
  
  export interface Gini {
  
    "2018": number;
  
  }
  
  
  
  export interface Idd {
  
    root: string;
  
    suffixes: string[];
  
  }
  
  
  
  export interface Languages {
  
    fin: string;
  
    swe: string;
  
  }
  
  
  
  export interface Maps {
  
    googleMaps: string;
  
    openStreetMaps: string;
  
  }
  
  
  
  export interface Name {
  
    common: string;
  
    official: string;
  
    nativeName: NativeName;
  
  }
  
  
  
  export interface NativeName {
  
    fin: Translation;
  
    swe: Translation;
  
  }
  
  
  
  export interface Translation {
  
    official: string;
  
    common: string;
  
  }
  
  
  
  export interface PostalCode {
  
    format: string;
  
    regex: string;
  
  }